pretty big commit adding support for skinit
skinit will run successfully

most of the hypervisor seems to be happy afterwards

however there seem to be some implicit assumptions about globals being initialized to 0, and possible also stack layout.  static variables may also be grouped into what I just called 'globals'

so right now the global flag skinit_status_flag is set to 1.  This makes the code think skinit already happened, and the execution should be identical to EMHF before I touched anything.

setting skinit_status_flag to 0 will then cause the new code to execute.  something goes wrong when it's time to startup the HVM.  I'm not sure what it is yet.  However, I've made a lot of changes, and some of them are somewhat delicate, so I'm nervous about having not committed in a while.

I intend to keep working on this until it can just be turned on by default.

jumped through some hoops to minimize the diff since there were some dos/unix issues.  I *think* this is everything and that it should just work.  Because of these issues the formatting is likely not ideal / atrocious.  However the diff should only contain relevant lines, and nothing to do with whitespace, etc.




git-svn-id: https://sparrow.ece.cmu.edu/svn/hyp@78 1c601a1c-44f4-4d93-90ac-4b632a93592e