skinit disabled in globals.h but is working
found a few flags in VM_CR that were not correct post-skinit; they're being fixed up in cstartup now

relocated SLB to 0x80000 (512K) and we get a lot further during boot now

seems the problem is likely the DEV protections put in place by SKINIT are preventing Windows from doing DMA that is usually does during bootup

We can either disable the DEV or move the SLB out of the way somewhere.  Leaning towards trying to stick it near the top of physical memory.  Or at least making it part of the hypervisor's reserved area.  It will need to be unity-mapped though





git-svn-id: https://sparrow.ece.cmu.edu/svn/hyp@83 1c601a1c-44f4-4d93-90ac-4b632a93592e