Use incremental linking instead of nested libraries
I had been building libraries (.a) in subdirectories, and then ar'ing
those into bigger libraries at lower levels. Unfortunately this doesn't
seem to work as expected. The nested library is still a .a archive inside
the base library; ar doesn't flatten them out. When trying to link against
the base library, the contents of the nested .a's is ignored.

So, instead we use incremental linking (ld -r) to link together a bunch
of .o's in a subdirectory into one 'o that includes all of their contents.
I think this is also the approach of the linux kernel...

git-svn-id: https://sparrow.ece.cmu.edu/svn/hyp@705 1c601a1c-44f4-4d93-90ac-4b632a93592e