beginning of encodeMemoryReference
Encoding a memory reference is a bit tricky. In an ideal situation, the
memory area will be mapped into the secure world at the same address.
However, it could be mapped at a different address, or could end up
being copied. This is up to the back-end driver, so the driver needs
a way to specify to the marshalling layer what address to actually
give when the shared memory reference is decoded.

Ultimately I ended up marshalling the reference with the address in the
caller's address space, with the marshalling code passing back the
offset into the decode buffer where that pointer is stored. The driver's
callback is then called and provided with that offset. If the addresses
will be unity mapped in the secure environment, the driver doesn't need
to do anything. If it will be mapped to a different address, which the
driver knows at that point, the driver can immediately rewrite the
address in the decode-buffer. If the driver needs to rewrite the address,
but doesn't know the remapping yet (for example, if the secure environment
is on a remote machine, so the driver may not know yet exactly where the
memory range will end up being copied), it can record the offset for
later rewriting.

git-svn-id: https://sparrow.ece.cmu.edu/svn/hyp@738 1c601a1c-44f4-4d93-90ac-4b632a93592e