Merge branch shared_mem
This adds the VMM command SHARE, which designates
a memory range to be passed to a PAL the next time
it is invoked. After SHAREing a memory range, the
application can no longer touch that range until
the next time the PAL returns, at which point
the range is no longer shared.



git-svn-id: https://sparrow.ece.cmu.edu/svn/hyp@903 1c601a1c-44f4-4d93-90ac-4b632a93592e