Include TPM_PCR_SELECTION in plaintext to be Sealed.
We had forgetten to include the TPM_PCR_SELECTION in a non-hashed form in the encrypted data,
so that it was impossible to know which PCRs were supposed to have been included.

Also fixed a bug where the IV was used twice: once as the IV and then again as the first 16 bytes of plaintext.

git-svn-id: https://sparrow.ece.cmu.edu/svn/hyp@1508 1c601a1c-44f4-4d93-90ac-4b632a93592e