Fixed decryption bug.  Suspect problem related to attempting to encrypt in-place.
This bug may be the explaination why the previous implementation of Seal opted for a "confounder" instead of an IV, since the in-place decrypt always caused the first block of decrypted plaintext to look like random data.

I didn't go the extra mile to guarantee that was the cause.  Solved it by dynamically allocating a "plaintext" buffer to hold everything that gets encrypted (lots more than just the input data; see comments in the code).

git-svn-id: https://sparrow.ece.cmu.edu/svn/hyp@1510 1c601a1c-44f4-4d93-90ac-4b632a93592e