tweaks to encode\decode-with-format-string interfaces
I realized that using %u for uint32 won't work correctly on 64-bit systems.
The correct way to get a format specifier for uint32 and uint64 is to
use the macros PRIu32 and PRIu64.

Previously in calling code I was using the literal format specifiers
in the format string rather than going through the TZI_EFMT_* macros.
However, since the calling code will need to use a macro (e.g. PRIu32)
for portability, they might as well use TZI format-specifier macros
directly.

To support this, I renamed the macros to be more succinct.

I took the leading '%' out of the macros to be consistent with how, e.g., PRIu32
macros are used.

I also renamed the functions to, e.g., TZIEncodeF instead of TZIEncodeFormat,
again for brevity.

git-svn-id: https://sparrow.ece.cmu.edu/svn/hyp@1537 1c601a1c-44f4-4d93-90ac-4b632a93592e