refactored scode_* where * is a utpm operation into a new file: pal_utpm*
Intention is that TrustVisor needs to contain some "glue" code to connect
hypercalls from individual PALs to the uTPM implementation that goes with that PAL.
However, the utpm.[ch] implementation should be sufficiently standalone as
to eventually support independent testing.

Thus, the purpose of pal_utpm.[ch] is to be the plumbing that connects
hypercalls in PALs to the right uTPM operation(s).

git-svn-id: https://sparrow.ece.cmu.edu/svn/hyp@2006 1c601a1c-44f4-4d93-90ac-4b632a93592e