Enhance uTPM Quote to return a TPM_PCR_COMPOSITE as an additional output parameter.
This functionality mirrors that of the hardware TPM and avoids a race condition where
the PCR values change between a PCR Read and a Quote, or a Quote then a Read.

This also simplifies verification of the quote since all necessary data is included
in one of (1) the input parameters to the Quote command, (2) output parameters from
the Quote command, or (3) well-known constants.

This builds but is untested.  Both the 'test' and 'attestation' PALs will need to
be updated to work with the additional parameters.

git-svn-id: https://sparrow.ece.cmu.edu/svn/hyp@2021 1c601a1c-44f4-4d93-90ac-4b632a93592e