modify openssl to use TERMIOS instead of TERMIO on linux platform.
Didn't want to implement termio in newlib. So far I've been implementing
all the interfaces according to the posix standard, but termio is
legacy and non-standardized. termios is the posix standardized
version.

The current solution of modifying openssl to choose termios on the linux platform is a bit
ugly for a couple reasons. First, it might break a 'true' linux build. Second, it
assumes we're building on linux.

The right solution is probably to add 'tsvc' as a first-class platform
in openssl and modify it appropriately throughout. This would be a bit
of an undertaking though, so I'll stick with the current approach of
using a slightly tweaked linux configuration instead.

git-svn-id: https://sparrow.ece.cmu.edu/svn/hyp@2116 1c601a1c-44f4-4d93-90ac-4b632a93592e