type-safe conversions between uint32_t and size_t
When linking against newlib, size_t and uint32_t are not identical,
and hence we get warnings and errors where we give one and the compiler
expects the other.

Note that when passing a pointer to size_t or uint32_t it is unsafe
to simply cast to a pointer of the other type, since they can be different
sizes. (and *will* be on 64-bit)

git-svn-id: https://sparrow.ece.cmu.edu/svn/hyp@2142 1c601a1c-44f4-4d93-90ac-4b632a93592e