Updated Makefile.in so that TrustVisor does build with these changes.  Tested and working (bootable, etc).
Note that the INSTALLDIR variable and associated activity are redundant with respect to existing functionality from automake, etc.

Right now, the files that get created in INSTALLDIR (currently CURDIR) are not used for anything when the Debian package gets created.
It reaches directly into emhfcore/x86/ and grabs the relevant files.

I think the easiest thing to do is teach 'make install' to use the files from INSTALLDIR.  that way, at least the dependencies will be linear.

git-svn-id: https://sparrow.ece.cmu.edu/svn/hyp@2470 1c601a1c-44f4-4d93-90ac-4b632a93592e