Updated Makefile.in to depend on INSTALLDIR. Comment to explain INSTALLDIR's relationship to core.
This causes all of the dependencies between created files to be linear.

emhfcore creates runtime.bin, sl.bin, and init.bin in its own tree.
emhfcore Makefile causes hypervisor-x86.bin.gz to get created in emhfcore/x86
emhfcore then uses INSTALLDIR as set in Makefile.in to copy init-x86.bin and hypervisor-x86.bin.gz to INSTALLDIR and prepend $(TARGET), which for us is 'trustvisor'.
This results in the creation of trustvisor-hypervisor-x86.bin.gz and trustvisor-init.bin in INSTALLDIR, which is presently CURDIR, which is trustvisor/trunk/code.

Makefile.in now copies those two files when invoked with make install-bin, instead of reaching into emhfcore/x86 directly.

Phew.

git-svn-id: https://sparrow.ece.cmu.edu/svn/hyp@2471 1c601a1c-44f4-4d93-90ac-4b632a93592e