New function validate_nv_access_controls() to confirm that an NV index has appropriate Locality and PCR restrictions in place.
Compiles but untested.  Right now we aren't rigorous about PCRs 17 and 18, and instead use 11 and 12 for testing purposes.
This represents a serious vulnerability but until we have an upgrade system for sealed storage being rigorous here
unnecessarily complicates debugging and development.

Debug print-outs complain loudly that the current configuration is vulnerable.

git-svn-id: https://sparrow.ece.cmu.edu/svn/hyp@2475 1c601a1c-44f4-4d93-90ac-4b632a93592e