build pal objects in a different directory
Source files that are used in both the pal and the regular program
need to be compiled twice. Before I was working around this with
custom rules for each such file, but this was getting unwieldy.
Instead we build all pal objects in a different sub-directory,
using the vpath directive to look for .c files in the main source
directory. Initially I tried using VPATH which looks for *all*
dependencies in the specified directory, but then additional
steps need to be taken to prevent it from using
object files from the 'regular program' build.

git-svn-id: https://sparrow.ece.cmu.edu/svn/hyp@2624 1c601a1c-44f4-4d93-90ac-4b632a93592e