fix hpt test by including _hpt.h directly
A design goal for hpt.h (now _hpt.h) is for it to be very modular,
and to *not* reference any global state in emhf. It's intended
as sort of a mini-library that is useful to both the core and to apps.

Trying to change the hpt test to include emhf.h instead gets hairy,
because that ends up referencing a lot of the emhfcore global state.

git-svn-id: https://sparrow.ece.cmu.edu/svn/hyp@2910 1c601a1c-44f4-4d93-90ac-4b632a93592e