collapsed back into one Makefile
I realized that a more straight-forward mechanism for getting make to
build separate object files for the pal was to create a new pattern
rule mapping source files in the root source directory to object
files in the pal-build directory.

I merged back the two makefiles and did a few other minor fixups.

git-svn-id: https://sparrow.ece.cmu.edu/svn/hyp@2934 1c601a1c-44f4-4d93-90ac-4b632a93592e