replace incorrect emhf_memprot_(get|set)_current_root_pagemap_address with emhf_memprot_(get|set)_(h_cr3|EPTP)
emhf_memprot_(get|set)_current_root_pagemap_address was implemented as if h_cr3 and EPTP contained
only the pointer to the root pagemap. However, they also contain additional fields that need to be masked
off when getting the root pagemap address, and preserved when setting the root pagemap address.

rather than duplicating the logic from the hpt module for the structure of those registers, in emhf_memprot
we'll just provide "raw" access to those registers, abstract logic for accessing that logic in hpt, and
some convenience composition functions in hpt_emhf

git-svn-id: https://sparrow.ece.cmu.edu/svn/hyp@4197 1c601a1c-44f4-4d93-90ac-4b632a93592e