undefine __CYGWIN__ when building newlib.
Otherwise newlib tries to pull in headers from /usr/include/cygwin, which
in turn try to pull in other system headers. Not sure why it does this,
but I think it might be becasue Cygwin uses newlib natively. The newlib
codebase might not differentiate correctly between "building on cygwin"
and "building for cygwin".

git-svn-id: https://sparrow.ece.cmu.edu/svn/hyp@4885 1c601a1c-44f4-4d93-90ac-4b632a93592e