replacement posix_memalign for windows
I'm not sure whether this is the best way to handle the problem since
we're building a *library*. autoconf has the macro AC_REPLACE_FUNCS
that does this in a more streamlined way for when you're building an
executable. It's unclear whether we actually ought to be including
posix_memalign in the library we build, though.

In principle we could run into trouble if an executable tries to link both
this library and another library B that employed the same strategy and has
its own posix_memalign replacement. library B could end up using our
posix_memalign replacement instead of its own, or vice versa. In this case
it should be OK since our implementation is a (nearly) conformant
implementation. In the case of partially implemented replacements it seems
you could run into trouble this way, though.

git-svn-id: https://sparrow.ece.cmu.edu/svn/hyp@5336 1c601a1c-44f4-4d93-90ac-4b632a93592e