remove version numbers from third party lib paths.
The version number it was originally derived from is preserved in history.
At some point we'll pull in newer versions from the canonical repos, at
which point the version number in the path will become inaccurate.
Rather than having to choose down the road whether to leave it be inaccurate
or break the paths in other projects, we'll just take off the version name
now.