ticket-28: xmhf-core: fix bug in read_segreg_cs; this caused the hypervisor to use the data segment selector instead of the code segment selector during intercept handling
how were we working despite this? think descriptor cache ;)
of course things start falling apart when we handle exceptions
(NMI) which push code selectors on stack in addition to RIP. This
causes the data segment selector to be pushed on stack (instead
of the code selector). Upon an IRETD from the NMI handler triggers
we therefore end up with a #GPF

tested extensively with the quiesce example hypapp and it
works successfully now!