generate index.html from any README.md
In the sourceforge dynamic markdown renderer, it makes sense to link
to a project subdirectory itself when talking about a sub-project,
because the directory listing view shows both the directory listing
and the rendered version of any README.md it finds there.

In offline viewing, they'll just get the directory listing. It may
not be obvious to them to click on the generated README.md.html.

So, we create index.html from any README.md. Truly offline viewing
will still give a directory listing, but it may be more obvious
that index.html is the intended target. When viewed through a web
server, the server should redirect to index.html automatically.